LayoutTests/http/tests/notifications/permission/permission-status-onchange-event-expected.txt

This test checks that the Permissions API on-change event works.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


No permission state is set
PASS permissionStatusState is "prompt"
Permission has been granted
PASS permissionStatusState is "granted"
PASS onChangeCalled became true
PASS successfullyParsed is true

TEST COMPLETE
